decompose timelinescreen into several files
