+++ frontend/src/screens/main/TimelineScreen.tsx
@@ -4,48 +4,19 @@ import {
   RefreshControl,
   ScrollView,
   Text,
-  TextInput,
   TouchableOpacity,
   View,
 } from 'react-native';
 import { BackgroundGradient } from '../../components/BackgroundGradient';
-import { LogDateTimePicker } from '../../components/LogDateTimePicker';
+import { FoodCard } from '../../components/FoodCard';
+import { SymptomCard } from '../../components/SymptomCard';
 import { foodLogService, FoodLogResponse } from '../../api/food_log_service';
 import { symptomLogService, SymptomLogResponse } from '../../api/symptom_log_service';
 import { useBankStore } from '../../store/bankStore';
 import { useUIStore } from '../../store/uiStore';
+import { TimelineEntry, FoodTimelineEntry, SymptomTimelineEntry } from '../../types/timeline';
 
-// ─── unified entry ────────────────────────────────────────────────────────────
-
-type FoodTimelineEntry = {
-  type: 'food';
-  id: string;
-  name: string;
-  ingredients: string[];
-  quantity: string | null;
-  notes: string | null;
-  timestamp: Date;
-};
-
-type SymptomTimelineEntry = {
-  type: 'symptom';
-  id: string;
-  name: string;
-  location: string;
-  sensation: string;
-  intensity: number;
-  duration: number | null;
-  notes: string | null;
-  timestamp: Date;
-};
-
-type TimelineEntry = FoodTimelineEntry | SymptomTimelineEntry;
-
-// ─── helpers ──────────────────────────────────────────────────────────────────
-
-function formatTime(date: Date): string {
-  return date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
-}
+// ─── screen-local helpers ─────────────────────────────────────────────────────
 
 function formatDateHeader(date: Date): string {
   const today = new Date();
@@ -71,356 +42,7 @@ function groupByDate(entries: TimelineEntry[]): { dateKey: string; label: string
   return Array.from(map.entries()).map(([dateKey, val]) => ({ dateKey, ...val }));
 }
 
-/** Maps a 1–10 intensity level to a full Tailwind bg class. */
-function intensityBgCls(level: number) {
-  if (level >= 7) return 'bg-remetra-rose';
-  if (level >= 4) return 'bg-remetra-coral';
-  return 'bg-remetra-mauve';
-}
-
-/** Maps a 1–10 intensity level to its hex value (needed for borderLeftColor). */
-function intensityHex(level: number) {
-  if (level >= 7) return '#C85A4A'; /* remetra-rose */
-  if (level >= 4) return '#D9806E'; /* remetra-coral */
-  return '#b2939b';                 /* remetra-mauve */
-}
-
-// ─── shared edit helpers ──────────────────────────────────────────────────────
-
-function EditLabel({ children }: { children: string }) {
-  return (
-    <Text className="text-xs font-semibold text-neutral-500 mb-1 mt-2.5 tracking-wide">
-      {children.toUpperCase()}
-    </Text>
-  );
-}
-
-function EditTextInput({
-  value,
-  onChange,
-  placeholder,
-  multiline,
-  integerOnly,
-}: {
-  value: string;
-  onChange: (v: string) => void;
-  placeholder?: string;
-  multiline?: boolean;
-  integerOnly?: boolean;
-}) {
-  const handleChange = (v: string) => {
-    onChange(integerOnly ? v.replace(/[^0-9]/g, '') : v);
-  };
-
-  return (
-    <TextInput
-      value={value}
-      onChangeText={handleChange}
-      placeholder={placeholder}
-      placeholderTextColor="#bbb"
-      multiline={multiline}
-      keyboardType={integerOnly ? 'number-pad' : 'default'}
-      className="border border-[#E0D0D8] rounded-lg px-2.5 py-2 text-sm text-neutral-700 bg-white"
-      style={{
-        minHeight: multiline ? 60 : undefined,
-        textAlignVertical: multiline ? 'top' : undefined,
-      }}
-    />
-  );
-}
-
-function SaveCancelRow({
-  saving,
-  onSave,
-  onCancel,
-}: {
-  saving: boolean;
-  onSave: () => void;
-  onCancel: () => void;
-}) {
-  return (
-    <View className="flex-row gap-2 mt-3.5">
-      <TouchableOpacity
-        onPress={onCancel}
-        className="flex-1 py-2.5 rounded-lg border border-remetra-border items-center"
-      >
-        <Text className="text-sm text-remetra-muted">Cancel</Text>
-      </TouchableOpacity>
-      <TouchableOpacity
-        onPress={onSave}
-        disabled={saving}
-        className="flex-[2] py-2.5 rounded-lg bg-remetra-burgundy items-center"
-        style={{ opacity: saving ? 0.6 : 1 }}
-      >
-        {saving ? (
-          <ActivityIndicator size="small" color="white" />
-        ) : (
-          <Text className="text-sm text-white font-semibold">Save Changes</Text>
-        )}
-      </TouchableOpacity>
-    </View>
-  );
-}
-
-// ─── food card ────────────────────────────────────────────────────────────────
-
-function FoodCard({
-  entry,
-  onUpdate,
-}: {
-  entry: FoodTimelineEntry;
-  onUpdate: (updated: Partial<FoodTimelineEntry>) => void;
-}) {
-  const [expanded, setExpanded] = useState(false);
-  const [quantity, setQuantity] = useState(entry.quantity ?? '');
-  const [notes, setNotes] = useState(entry.notes ?? '');
-  const [timestamp, setTimestamp] = useState(entry.timestamp);
-  const [saving, setSaving] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const handleCancel = () => {
-    setQuantity(entry.quantity ?? '');
-    setNotes(entry.notes ?? '');
-    setTimestamp(entry.timestamp);
-    setError(null);
-    setExpanded(false);
-  };
-
-  const handleSave = async () => {
-    setSaving(true);
-    setError(null);
-    try {
-      await foodLogService.updateFoodLog(entry.id, {
-        quantity: quantity || undefined,
-        notes: notes || undefined,
-        timestamp: timestamp.toISOString(),
-      });
-      onUpdate({ quantity: quantity || null, notes: notes || null, timestamp });
-      setExpanded(false);
-    } catch {
-      setError('Failed to save. Please try again.');
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  return (
-    <View
-      className="rounded-xl mb-2.5 overflow-hidden"
-      style={{ backgroundColor: '#FEF0E7', borderLeftWidth: 4, borderLeftColor: '#fca450' /* remetra-orange */ }}
-    >
-      <TouchableOpacity
-        onPress={() => setExpanded((prev) => !prev)}
-        activeOpacity={0.7}
-        className="p-3.5"
-      >
-        <View className="flex-row justify-between items-start">
-          <View className="flex-1">
-            <View className="flex-row items-center gap-1.5 mb-1">
-              <Text className="text-base">🍽</Text>
-              <Text style={{ fontSize: 16, fontWeight: '600', color: '#7C4A1E' }}>{entry.name}</Text>
-            </View>
-            {entry.quantity ? (
-              <Text style={{ fontSize: 13, color: '#A0673A', marginBottom: 2 }}>Qty: {entry.quantity}</Text>
-            ) : null}
-            {entry.ingredients.length > 0 && (
-              <Text style={{ fontSize: 12, color: '#B07850' }} numberOfLines={expanded ? undefined : 2}>
-                {entry.ingredients.join(', ')}
-              </Text>
-            )}
-            {!expanded && entry.notes ? (
-              <Text style={{ fontSize: 12, color: '#9E7050', marginTop: 4, fontStyle: 'italic' }} numberOfLines={1}>
-                {entry.notes}
-              </Text>
-            ) : null}
-          </View>
-          <View className="items-end gap-1 ml-2">
-            <Text style={{ fontSize: 12, color: '#B07850' }}>{formatTime(entry.timestamp)}</Text>
-            <Text style={{ fontSize: 11, color: '#B07850' }}>{expanded ? '▲ collapse' : '▼ edit'}</Text>
-          </View>
-        </View>
-      </TouchableOpacity>
-
-      {expanded && (
-        <View className="px-3.5 pb-3.5" style={{ borderTopWidth: 1, borderTopColor: '#FAD9C4' }}>
-          <EditLabel>Quantity</EditLabel>
-          <EditTextInput value={quantity} onChange={setQuantity} placeholder="e.g. 2" integerOnly />
-
-          <EditLabel>Date & Time</EditLabel>
-          <LogDateTimePicker value={timestamp} onChange={setTimestamp} />
-
-          <EditLabel>Notes</EditLabel>
-          <EditTextInput value={notes} onChange={setNotes} placeholder="Optional notes…" multiline />
-
-          {error ? (
-            <Text className="text-remetra-rose text-xs mt-2">{error}</Text>
-          ) : null}
-
-          <SaveCancelRow saving={saving} onSave={handleSave} onCancel={handleCancel} />
-        </View>
-      )}
-    </View>
-  );
-}
-
-// ─── symptom card ─────────────────────────────────────────────────────────────
-
-function SymptomCard({
-  entry,
-  onUpdate,
-}: {
-  entry: SymptomTimelineEntry;
-  onUpdate: (updated: Partial<SymptomTimelineEntry>) => void;
-}) {
-  const [expanded, setExpanded] = useState(false);
-  const [intensity, setIntensity] = useState(entry.intensity);
-  const [duration, setDuration] = useState(entry.duration?.toString() ?? '');
-  const [hasDuration, setHasDuration] = useState(entry.duration !== null);
-  const [notes, setNotes] = useState(entry.notes ?? '');
-  const [timestamp, setTimestamp] = useState(entry.timestamp);
-  const [saving, setSaving] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const handleCancel = () => {
-    setIntensity(entry.intensity);
-    setDuration(entry.duration?.toString() ?? '');
-    setHasDuration(entry.duration !== null);
-    setNotes(entry.notes ?? '');
-    setTimestamp(entry.timestamp);
-    setError(null);
-    setExpanded(false);
-  };
-
-  const handleSave = async () => {
-    setSaving(true);
-    setError(null);
-    try {
-      const durationVal = hasDuration && duration ? parseInt(duration, 10) : undefined;
-      await symptomLogService.updateSymptomLog(entry.id, {
-        intensity,
-        timestamp: timestamp.toISOString(),
-        duration: durationVal,
-        notes: notes || undefined,
-      });
-      onUpdate({
-        intensity,
-        timestamp,
-        duration: durationVal ?? null,
-        notes: notes || null,
-      });
-      setExpanded(false);
-    } catch {
-      setError('Failed to save. Please try again.');
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  const intHex = intensityHex(intensity);
-  const intBgCls = intensityBgCls(intensity);
-
-  return (
-    <View
-      className="rounded-xl mb-2.5 overflow-hidden"
-      style={{ backgroundColor: '#FBF0F2', borderLeftWidth: 4, borderLeftColor: intHex }}
-    >
-      <TouchableOpacity
-        onPress={() => setExpanded((prev) => !prev)}
-        activeOpacity={0.7}
-        className="p-3.5"
-      >
-        <View className="flex-row justify-between items-start">
-          <View className="flex-1">
-            <View className="flex-row items-center gap-1.5 mb-1">
-              <Text className="text-base">🩺</Text>
-              <Text style={{ fontSize: 16, fontWeight: '600', color: '#7B3B4E' }}>{entry.name}</Text>
-            </View>
-            <View className="flex-row items-center gap-2 mb-0.5">
-              <View className={`${intBgCls} rounded-[10px] px-2 py-0.5`}>
-                <Text className="text-white text-xs font-semibold">
-                  Intensity {entry.intensity}/10
-                </Text>
-              </View>
-              {entry.duration ? (
-                <Text style={{ fontSize: 12, color: '#9E7080' }}>{entry.duration} min</Text>
-              ) : null}
-            </View>
-            {entry.location ? (
-              <Text style={{ fontSize: 12, color: '#9E7080' }}>Location: {entry.location}</Text>
-            ) : null}
-            {!expanded && entry.notes ? (
-              <Text style={{ fontSize: 12, color: '#9E7080', marginTop: 4, fontStyle: 'italic' }} numberOfLines={1}>
-                {entry.notes}
-              </Text>
-            ) : null}
-          </View>
-          <View className="items-end gap-1 ml-2">
-            <Text style={{ fontSize: 12, color: '#9E7080' }}>{formatTime(entry.timestamp)}</Text>
-            <Text style={{ fontSize: 11, color: '#9E7080' }}>{expanded ? '▲ collapse' : '▼ edit'}</Text>
-          </View>
-        </View>
-      </TouchableOpacity>
-
-      {expanded && (
-        <View className="px-3.5 pb-3.5" style={{ borderTopWidth: 1, borderTopColor: '#F0D0DA' }}>
-          <EditLabel>Intensity</EditLabel>
-          <View className="flex-row flex-wrap gap-1.5">
-            {Array.from({ length: 10 }, (_, i) => i + 1).map((n) => (
-              <TouchableOpacity
-                key={n}
-                onPress={() => setIntensity(n)}
-                className={`w-9 h-9 rounded-[18px] justify-center items-center ${
-                  intensity === n ? intensityBgCls(n) : 'bg-[#F5E8EC]'
-                }`}
-                style={{
-                  borderWidth: intensity === n ? 0 : 1,
-                  borderColor: '#E0C8D0',
-                }}
-              >
-                <Text
-                  className={`text-[13px] font-semibold ${intensity === n ? 'text-white' : 'text-[#9E7080]'}`}
-                >
-                  {n}
-                </Text>
-              </TouchableOpacity>
-            ))}
-          </View>
-
-          <EditLabel>Date & Time</EditLabel>
-          <LogDateTimePicker value={timestamp} onChange={setTimestamp} />
-
-          <View className="flex-row items-center gap-2 mt-2.5">
-            <TouchableOpacity
-              onPress={() => setHasDuration((prev) => !prev)}
-              className={`w-5 h-5 rounded border justify-center items-center ${
-                hasDuration ? 'bg-remetra-burgundy border-remetra-burgundy' : 'bg-white border-remetra-border'
-              }`}
-            >
-              {hasDuration ? <Text className="text-white text-xs">✓</Text> : null}
-            </TouchableOpacity>
-            <Text style={{ fontSize: 13, color: '#7B3B4E' }}>Duration (minutes)</Text>
-          </View>
-          {hasDuration && (
-            <View className="mt-1.5">
-              <EditTextInput value={duration} onChange={setDuration} placeholder="e.g. 30" integerOnly />
-            </View>
-          )}
-
-          <EditLabel>Notes</EditLabel>
-          <EditTextInput value={notes} onChange={setNotes} placeholder="Optional notes…" multiline />
-
-          {error ? (
-            <Text className="text-remetra-rose text-xs mt-2">{error}</Text>
-          ) : null}
-
-          <SaveCancelRow saving={saving} onSave={handleSave} onCancel={handleCancel} />
-        </View>
-      )}
-    </View>
-  );
-}
-
-// ─── main screen ──────────────────────────────────────────────────────────────
+// ─── screen ───────────────────────────────────────────────────────────────────
 
 export function TimelineScreen() {
   const { foods, symptoms, fetchFoods, fetchSymptoms } = useBankStore();
